# apps/desktop/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eigenheim engine",
    "version": "0.1.0"
  },
  "paths": {
...
    "/logic": {
      "get": {
        "summary": "Logic",
        "operationId": "logic_logic_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogicOut"
                  },
                  "title": "Response Logic Logic Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Logic",
        "operationId": "create_logic_logic_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogicIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogicCreateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/definitions": {
      "get": {
        "summary": "List Definitions",
        "description": "List the canonical versioned definition of every Logic metric.",
        "operationId": "list_definitions_definitions_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogicDefinitionListItem"
                  },
                  "title": "Response List Definitions Definitions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/definitions/{logic_id}": {
      "get": {
        "summary": "Get Definition",
        "description": "Get the full versioned definition of one Logic metric.",
        "operationId": "get_definition_definitions__logic_id__get",
        "parameters": [
          {
            "name": "logic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Logic Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogicDefinitionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/logic/{logic_id}/drafts": {
      "get": {
        "summary": "List Logic Drafts",
        "description": "List all pending draft versions for a Logic metric.",
        "operationId": "list_logic_drafts_logic__logic_id__drafts_get",
        "parameters": [
          {
            "name": "logic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Logic Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DraftOut"
                  },
                  "title": "Response List Logic Drafts Logic  Logic Id  Drafts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/logic/drafts/all": {
      "get": {
        "summary": "List All Drafts",
        "description": "List all pending draft versions across every Logic metric.",
        "operationId": "list_all_drafts_logic_drafts_all_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DraftOut"
                  },
                  "title": "Response List All Drafts Logic Drafts All Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/logic/{logic_id}/versions/{version}/promote": {
      "post": {
        "summary": "Promote Version",
        "description": "Promote a draft version of a Logic metric to live (human gate).",
        "operationId": "promote_version_logic__logic_id__versions__version__promote_post",
        "parameters": [
          {
            "name": "logic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Logic Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromoteRejectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/logic/{logic_id}/versions/{version}/reject": {
      "post": {
        "summary": "Reject Version",
        "description": "Reject a draft version of a Logic metric (human gate).",
        "operationId": "reject_version_logic__logic_id__versions__version__reject_post",
        "parameters": [
          {
            "name": "logic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Logic Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromoteRejectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/logic/{logic_id}/audit": {
      "get": {
        "summary": "Get Logic Audit",
        "description": "Return the hash-chain audit trail for one Logic metric.",
        "operationId": "get_logic_audit_logic__logic_id__audit_get",
        "parameters": [
          {
            "name": "logic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Logic Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEntryOut"
                  },
                  "title": "Response Get Logic Audit Logic  Logic Id  Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuditEntryOut": {
        "properties": {
          "logic_id": {
            "type": "string",
            "title": "Logic Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "hash": {
            "type": "string",
            "title": "Hash"
          },
          "prev_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev Hash"
          },
          "at": {
            "type": "string",
            "title": "At"
          }
        },
        "type": "object",
        "required": [
          "logic_id",
          "version",
          "action",
          "actor",
          "hash",
          "at"
        ],
        "title": "AuditEntryOut"
      },
...
      "DraftOut": {
        "properties": {
          "logic_id": {
            "type": "string",
            "title": "Logic Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "fmt": {
            "type": "string",
            "title": "Fmt"
          },
          "inputs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Inputs"
          },
          "expression": {
            "type": "string",
            "title": "Expression"
          },
          "validated": {
            "type": "string",
            "title": "Validated"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          }
        },
        "type": "object",
        "required": [
          "logic_id",
          "version",
          "name",
          "description",
          "fmt",
          "inputs",
          "expression",
          "validated",
          "created_at",
          "status"
        ],
        "title": "DraftOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InputIn": {
        "properties": {
          "alias": {
            "type": "string",
            "title": "Alias"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "alias",
          "kind",
          "params"
        ],
        "title": "InputIn"
      },
...
      "LogicCreateOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "expression": {
            "type": "string",
            "title": "Expression"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "expression"
        ],
        "title": "LogicCreateOut"
      },
      "LogicDefinitionListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "expression": {
            "type": "string",
            "title": "Expression"
          },
          "inputs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Inputs"
          },
          "fmt": {
            "type": "string",
            "title": "Fmt"
          },
          "validated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validated"
          },
          "sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha"
          },
          "pending_drafts": {
            "type": "integer",
            "title": "Pending Drafts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "version",
          "expression",
          "inputs",
          "fmt",
          "pending_drafts"
        ],
        "title": "LogicDefinitionListItem",
        "description": "Lightweight definition as returned by GET /definitions (no history)."
      },
      "LogicDefinitionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "expression": {
            "type": "string",
            "title": "Expression"
          },
          "inputs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Inputs"
          },
          "fmt": {
            "type": "string",
            "title": "Fmt"
          },
          "validated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validated"
          },
          "sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha"
          },
          "pending_drafts": {
            "type": "integer",
            "title": "Pending Drafts"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/LogicVersionHistoryItem"
            },
            "type": "array",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "version",
          "expression",
          "inputs",
          "fmt",
          "pending_drafts",
          "history"
        ],
        "title": "LogicDefinitionOut",
        "description": "Full definition as returned by GET /definitions/{id} (includes history)."
      },
      "LogicIn": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "fmt": {
            "type": "string",
            "title": "Fmt",
            "default": "number"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/InputIn"
            },
            "type": "array",
            "title": "Inputs"
          },
          "expression": {
            "type": "string",
            "title": "Expression"
          },
          "validated": {
            "type": "string",
            "title": "Validated",
            "default": "\u0441\u0435\u0433\u043e\u0434\u043d\u044f"
          }
        },
        "type": "object",
        "required": [
          "name",
          "inputs",
          "expression"
        ],
        "title": "LogicIn"
      },
      "LogicOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "usage": {
            "type": "integer",
            "title": "Usage"
          },
          "expression": {
            "type": "string",
            "title": "Expression"
          },
          "badge": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Badge"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "usage",
          "expression"
        ],
        "title": "LogicOut"
      },
      "LogicVersionHistoryItem": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "expression": {
            "type": "string",
            "title": "Expression"
          },
          "validated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validated"
          },
          "sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha"
          },
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          }
        },
        "type": "object",
        "required": [
          "version",
          "status",
          "expression"
        ],
        "title": "LogicVersionHistoryItem"
      },
...
      "PromoteRejectOut": {
        "properties": {
          "logic_id": {
            "type": "string",
            "title": "Logic Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "spec_hash": {
            "type": "string",
            "title": "Spec Hash"
          },
          "audit_hash": {
            "type": "string",
            "title": "Audit Hash"
          }
        },
        "type": "object",
        "required": [
          "logic_id",
          "version",
          "status",
          "spec_hash",
          "audit_hash"
        ],
        "title": "PromoteRejectOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}